pyproject.toml
[project]
name = "synthetic-dataset-generation-suite"
version = "0.1.0"
description = "Multi-provider synthetic reasoning dataset generation pipeline"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0",
    "pyyaml",
    "click",
    "datasets",
    "arxiv",
    "semanticscholar",
    "pymupdf",
    "requests",
]

[project.optional-dependencies]
gpu = ["pynvml"]
dev = ["pytest", "ruff"]
web = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "sqlalchemy>=2.0",
    "pydantic>=2.0",
    "python-jose[cryptography]>=3.3",
    "bcrypt>=4.0",
    "cryptography>=41.0",
    "huggingface_hub>=0.20",
]

[tool.setuptools.packages.find]
include = ["sdgs*"]

[project.scripts]
sdgs = "sdgs.cli:cli"
